move some methods from back_api to service layer
