+++ backend/src/back_api/del_record.py
@@ -2,11 +2,11 @@ import logging
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, Request
+from services.token import get_current_user
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
 from back_api.schemas import Message, RemoveRecordRequest
-from back_api.token import get_current_user
 from database.database import get_session
 from database.hash import decrypt_id
 from repository.record import remove_record_row_by_id
