+++ backend/src/back_api/edit_record.py
@@ -7,10 +7,10 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
 from back_api.schemas import EditRecordRequest, Message
-from back_api.token import get_current_user
 from database.database import get_session
 from database.hash import decrypt_id
 from repository.record import edit_record_by_id
+from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
