+++ backend/src/back_api/get_publ.py
@@ -6,9 +6,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
 from back_api.schemas import PublResponse
-from back_api.token import get_current_user
 from database.database import get_session
 from repository.user import username_and_publication
+from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
