+++ backend/src/back_api/get_records_file.py
@@ -1,61 +1,19 @@
-import io
 import logging
-from collections.abc import Sequence
 from datetime import datetime
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, Request
 from fastapi.responses import StreamingResponse
-from openpyxl import Workbook
-from openpyxl.styles import Font
-from openpyxl.utils import get_column_letter
 from sqlalchemy.ext.asyncio import AsyncSession
-from starlette.responses import ContentStream
 
 from back_api.rate_limiter import limiter
-from back_api.token import get_current_user
 from database.database import get_session
-from database.models import Record
 from repository.record import get_user_records
+from service import repository
+from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
-
-COLUMN_MAPPING = {
-    "datetime": "Дата добавления записи",
-    "adm_country": "Страна",
-    "adm_region": "Регион",
-    "adm_district": "Район",
-    "adm_loc": "Место сбора",
-    "geo_nn": "Широта (десятич.)",
-    "geo_ee": "Долгота (десятич.)",
-    "geo_nn_raw": "Широта (изнач.)",
-    "geo_ee_raw": "Долгота (изнач.)",
-    "geo_origin": "Происхождение координат",
-    "geo_REM": "Примечания к расположению",
-    "eve_YY": "Год",
-    "eve_MM": "Месяц",
-    "eve_DD": "День",
-    "eve_day_def": "Определён ли день",
-    "eve_habitat": "Биотоп",
-    "eve_effort": "Выборочное усиление",
-    "abu_coll": "Коллектор",
-    "eve_REM": "Примечания к сбору материала",
-    "tax_fam": "Семейство",
-    "tax_gen": "Род",
-    "tax_sp": "Вид",
-    "tax_sp_def": "Определён ли вид",
-    "tax_nsp": "Описан ли как новый вид",
-    "type_status": "Типовой статус",
-    "tax_REM": "Таксономические примечания",
-    "abu": "Общее кол-во особей",
-    "abu_details": "Кол-во особей каждого пола/зрелости",
-    "abu_ind_rem": "Комментарий к особям",
-    "geo_uncert": "Радиус неточности координат, м",
-    "eve_YY_end": "Конечный год",
-    "eve_MM_end": "Конечный месяц",
-    "eve_DD_end": "Конечный день",
-}
 
 
 @router.post("/get_records_data")
@@ -83,7 +41,7 @@ async def get_records_data(
         }
 
         return StreamingResponse(
-            generate_excel(records),
+            repository.generate_excel(records),
             media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
             headers=headers,
         )
@@ -91,37 +49,3 @@ async def get_records_data(
     except Exception as e:
         logger.error(f"Exception in get_records_data: {str(e)}", exc_info=True)
         raise HTTPException(status_code=500, detail=str(e)) from e
-
-
-# NOTE: not sure if ContentStream is used correctly here, probably should check later
-def generate_excel(records: Sequence[Record]) -> ContentStream:
-    output = io.BytesIO()
-    wb = Workbook()
-    ws = wb.active
-
-    if ws is None:
-        logger.error("ws is None")
-        raise Exception
-
-    headers = [
-        COLUMN_MAPPING[field]
-        for field in COLUMN_MAPPING
-        if field not in ["id", "publ_id", "ip", "errors", "type", "adm_verbatim"]
-    ]
-    ws.append(headers)
-
-    for col in range(1, len(headers) + 1):
-        ws.column_dimensions[get_column_letter(col)].width = 20
-        ws.cell(row=1, column=col).font = Font(bold=True)
-
-    for record in records:
-        row = [
-            getattr(record, field)
-            for field in COLUMN_MAPPING
-            if field not in ["id", "publ_id", "ip", "errors", "type", "adm_verbatim"]
-        ]
-        ws.append(row)
-
-    wb.save(output)
-    output.seek(0)
-    yield output.read()
