+++ backend/src/back_api/next_publ.py
@@ -5,10 +5,10 @@ from fastapi import APIRouter, Depends, HTTPException, Request, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
-from back_api.token import get_current_user
 from database.database import get_session
 from repository.publication import is_publ_filled
 from repository.user import get_user, update_user
+from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
