+++ backend/src/back_api/publ_from_hash.py
@@ -6,10 +6,10 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from back_api.rate_limiter import limiter
 from back_api.schemas import PublResponse, RecordHashRequest
-from back_api.token import get_current_user
 from database.database import get_session
 from database.hash import decrypt_id
 from repository.record import publ_by_hash
+from service.token import get_current_user
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
