+++ backend/src/back_api/support.py
@@ -6,12 +6,12 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.status import HTTP_404_NOT_FOUND
 
-from back_api.messages import send_support_message
 from back_api.rate_limiter import limiter
 from back_api.schemas import Message, SupportRequest
 from back_api.util import get_http_session
 from database.database import get_session
 from repository.user import get_user_id_by_username
+from service.support import send_support_message
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
