+++ backend/src/back_api/user_image.py
@@ -1,4 +1,3 @@
-import io
 import logging
 from typing import Annotated
 
@@ -7,36 +6,10 @@ from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import StreamingResponse
 
 from back_api.util import get_http_session
-from config.config import BOT_TOKEN
+from service.telegram import fetch_telegram_photo
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
-
-
-async def fetch_telegram_photo(
-    session: aiohttp.ClientSession,
-    user_id: int,
-) -> io.BytesIO | None:
-    url = f"https://api.telegram.org/bot{BOT_TOKEN}/getUserProfilePhotos"
-    async with session.get(url, params={"user_id": user_id, "limit": 1}) as resp:
-        data = await resp.json()
-
-    photos = data.get("result", {}).get("photos", [])
-    if not photos:
-        return None
-
-    file_id = photos[0][-1]["file_id"]
-    file_info_url = f"https://api.telegram.org/bot{BOT_TOKEN}/getFile"
-    async with session.get(
-        file_info_url, params={"file_id": file_id}
-    ) as file_info_resp:
-        file_info = await file_info_resp.json()
-
-    file_path = file_info["result"]["file_path"]
-    file_url = f"https://api.telegram.org/file/bot{BOT_TOKEN}/{file_path}"
-    async with session.get(file_url) as file_resp:
-        # TODO: Streaming response?
-        return io.BytesIO(await file_resp.read())
 
 
 @router.get("/user_photo")
